+++ src/client/src/pixel.module.css
