+++ src/server/routes/api/alive.api.ts
