+++ src/server/routes/debug/stream.api.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express'
-import { dynamic } from '../../backend/shared'
+import { dynamic } from '../../../backend/shared'
 
 const connections = new Set<Response>()
